requirements.txt
torch
torchvision
fastapi
uvicorn
streamlit
pillow
numpy
pyyaml
requests
scikit-learn